% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_INE.R
\docType{data}
\name{INE_padron_10_17_muni}
\alias{INE_padron_10_17_muni}
\title{Poblacion de MUNICIPIOS PADRON INE (SPAIN) 2010-2017}
\format{A data frame with 64.948 registros and 8 variables:
\itemize{
\item INECodMuni: PANCRUDO = 44177
\item Code_Unidad_poblacional: Pancrudo = 44177000000
\item Nombre_Unidad_poblacional: PANCRUDO
\item Provincia: PANCRUDO esta en Teruel = 44
\item Municipio: Codigo INE municipio Pancrudo = 177
\item anyo: periodo
\item Poblacion_Total: Poblacion Total
\item Poblacion_H: Hombres
\item Poblacion_M: Mujeres
}}
\source{
\url{http://www.ine.es}
}
\usage{
INE_padron_10_17_muni
}
\description{
Poblacion de los MUNICIPIOS: PADRON INE (SPAIN) 2010-2017
64.948 registros
8 variables
}
\examples{
\dontrun{
 df <- INE_padron_10_17_muni
 df_2017 <- df \%>\% filter(anyo == 2017)
}

}
\keyword{datasets}
