% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_INE.R
\docType{data}
\name{INE_padron_10_17}
\alias{INE_padron_10_17}
\title{Poblacion PADRON INE (SPAIN) 210-2017}
\format{A data frame with 1.219.020 registros and 13 variables:
\itemize{
\item X1: Microdato original
\item anyo: periodo
\item Code_Unidad_poblacional: Pancrudo = 44177000000
\item Nombre_Unidad_poblacional: PANCRUDO
\item Poblacion_Total: Poblacion Total
\item Poblacion_H: Hombres
\item Poblacion_M: Mujeres
\item Provincia: PANCRUDO esta en Teruel = 44
\item Municipio: Codigo INE municipio Pancrudo = 177
\item Entidad_colectiva: codigo de 2 digitos
\item Entidad_singular: codigo de 2 digitos
\item Nucleo_o_diseminado: codigo de 2 digitos
\item TIPO:    Municipios, Entidad Singular, Nucleos ....
}}
\source{
\url{http://www.ine.es}
}
\usage{
INE_padron_10_17
}
\description{
CPoblacion PADRON INE (SPAIN) 210-2017
1.219.020 registros
13 variables
}
\examples{
\dontrun{
 df <- INE_padron_10_17
}

}
\keyword{datasets}
