% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanishRpoblacion-package.R
\docType{package}
\name{spanishRpoblacion-package}
\alias{spanishRpoblacion}
\alias{spanishRpoblacion-package}
\title{spanishRpoblacion: Proporciona datos de poblacion para entidades españolas (provenientes del padron INE)}
\description{
Datos de poblacion provenientes del Padron del INE
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/perezp44/spanishRpoblacion}
  \item Report bugs at \url{https://github.com/perezp44/spanishRpoblacion/issues}
}

}
\author{
\strong{Maintainer}: Pedro J. Perez \email{pedro.j.perez@uv.es}

}
\keyword{internal}
